+++ webpack.prod.js
@@ -1,7 +1,7 @@
 const { merge } = require('webpack-merge');
-const common = require("./webpack.common");
+const common = require('./webpack.common');
 
 module.exports = merge(common, {
   mode: 'production',
-  devtool: 'source-map',
+  devtool: 'source-map'
 });
